[S12P31B106-311] frontend 폴더 내 nodemodules 제거

See merge request s12-final/S12P31B106!306

+++ backend/src/main/java/com/voiceprint/backend/api/auth/dto/DiaryResponse.java
@@ -1,7 +1,6 @@
 package com.voiceprint.backend.api.auth.dto;
 
-import com.voiceprint.backend.domain.diary.Diary;
-import lombok.AllArgsConstructor;
+import com.voiceprint.backend.domain.Entity.Diary;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
@@ -11,7 +10,7 @@ import java.time.LocalDateTime;
 //@AllArgsConstructor
 @Slf4j
 public class DiaryResponse {
-    private Long diaryId;
+    private Integer diaryId;
     private String title;
     private String content;
     private LocalDateTime createdAt;
